requirements.txt
tensorflow-gpu==1.14.0
numpy==1.16.0
pandas==0.24.2
scikit-learn==0.20.3
scipy==1.2.1
tqdm==4.64.1
